+++ poem-openapi-derive/src/object.rs
@@ -8,7 +8,7 @@ use quote::quote;
 use syn::{ext::IdentExt, Attribute, DeriveInput, Error, GenericParam, Generics, Path, Type};
 
 use crate::{
-    common_args::{ConcreteType, DefaultValue, RenameRule, RenameRuleExt, RenameTarget},
+    common_args::{ConcreteType, DefaultValue, RenameRule, RenameRuleExt},
     error::GeneratorResult,
     utils::{get_crate_name, get_summary_and_description, optional_literal},
     validators::Validators,
@@ -79,10 +79,7 @@ pub(crate) fn generate(args: DeriveInput) -> GeneratorResult<TokenStream> {
             );
         }
     };
-    let oai_typename = args
-        .rename
-        .clone()
-        .unwrap_or_else(|| RenameTarget::Type.rename(ident.to_string()));
+    let oai_typename = args.rename.clone().unwrap_or_else(|| ident.to_string());
     let (title, description) = get_summary_and_description(&args.attrs)?;
     let mut deserialize_fields = Vec::new();
     let mut serialize_fields = Vec::new();
@@ -129,10 +126,10 @@ pub(crate) fn generate(args: DeriveInput) -> GeneratorResult<TokenStream> {
             .into());
         }
 
-        let field_name = field.rename.clone().unwrap_or_else(|| {
-            args.rename_all
-                .rename(field_ident.unraw().to_string(), RenameTarget::Field)
-        });
+        let field_name = field
+            .rename
+            .clone()
+            .unwrap_or_else(|| args.rename_all.rename(field_ident.unraw().to_string()));
         let (field_title, field_description) = get_summary_and_description(&field.attrs)?;
         let field_title = optional_literal(&field_title);
         let field_description = optional_literal(&field_description);